fix(backend): code cleanup after carlo migratio
