+++ lib/launcher.js
@@ -10,20 +10,15 @@ const path = require('path');
 const carlo = require('carlo');
 const { rpc, rpc_process } = require('carlo/rpc');
 
-const readline = require('readline');
-const {Readable} = require('stream');
 const removeFolder = require('rimraf');
-const {URL} = require('url');
 const util = require('util');
-const querystring = require('querystring');
 
-const {FileSystem} = require('./file_system_backend.js');
+const { Backend } = require('./backend.js');
 
 const fsCopyFile = fs.copyFile ? util.promisify(fs.copyFile) : function(source, target) {
   require('fs-copy-file-sync')(source, target);
   return Promise.resolve();
 };
-const fsReadFile = util.promisify(fs.readFile);
 const fsMkdir = util.promisify(fs.mkdir);
 const fsMkdtemp = util.promisify(fs.mkdtemp);
 const fsWriteFile = util.promisify(fs.writeFile);
@@ -34,15 +29,6 @@ async function launch(options) {
     bgcolor: '#242424',
     userDataDir
   });
-  app.safeEvaluate = async(...args) => {
-    try {
-      return await app.evaluate(...args);
-    } catch (e) {
-      if (!e.message.includes('Session closed.') && !e.message.includes('Target closed.'))
-        throw e;
-    }
-  };
-
   let pkg = null;
   try {
     pkg = require(path.join(options.cwd, 'package.json'));
@@ -50,7 +36,6 @@ async function launch(options) {
   }
   const preloadJs = path.join(userDataDir, 'preload.js');
   await fsCopyFile(require.resolve('../ndb-inspect/preload.js'), preloadJs);
-
   await Promise.all([
     app.exposeFunction('getProcessInfo', () => ({
       execPath: process.execPath,
@@ -63,16 +48,8 @@ async function launch(options) {
       preload: preloadJs,
       nddSharedStore: path.join(options.configDir, 'ndd_store'),
       configDir: options.configDir
-    })),
-    app.exposeFunction('bringToFront', () => app.mainWindow().bringToFront()),
-    app.exposeFunction('loadSourceMap', loadSourceMap)
+    }))
   ]);
-  app.on('exit', cleanupAndExit);
-
-  async function cleanupAndExit() {
-    if (!options.doNotProcessExit)
-      process.exit(0);
-  }
 
   const overridesFolder = options.debugFrontend
     ? path.dirname(require.resolve(`../front_end/${options.appName}.json`))
@@ -82,10 +59,13 @@ async function launch(options) {
   app.serveFolder(path.join(__dirname, '..'));
   app.serveFolder(path.dirname(require.resolve(`chrome-devtools-frontend/front_end/ndb_app.json`)));
 
+  const backend = new Backend(app);
+  app.on('exit', () => {
+    backend.dispose();
+    setTimeout(() => process.exit(0), 0);
+  });
   await app.load(`${options.appName}.html?experiments=true&debugFrontend=${options.debugFrontend}`,
-      rpc.handle({
-        createService: async(fullName, ...args) => await rpc_process.spawn(fullName, {args: args})
-      }));
+      rpc.handle(backend));
   return app;
 }
 
@@ -118,86 +98,6 @@ async function setupUserDataDir(configDirectory, doNotCopyPreferences) {
   });
 
   return userDataDir;
-}
-
-async function loadSourceMap(sourceMapURL, compiledURL) {
-  try {
-    let payload;
-    if (sourceMapURL.startsWith('data:')) {
-      const [metadata, ...other] = sourceMapURL.split(',');
-      const urlPayload = other.join(',');
-      const isBase64 = metadata.endsWith(';base64');
-      payload = JSON.parse(Buffer.from(isBase64 ? urlPayload : querystring.unescape(urlPayload), isBase64 ? 'base64' : 'utf8').toString('utf8'));
-    } else {
-      const fileURL = new URL(sourceMapURL);
-      const content = await fsReadFile(fileURL, 'utf8');
-      payload = JSON.parse(content);
-    }
-    await removeSourceContentIfMatch(sourceMapURL, compiledURL, payload);
-    return {payload};
-  } catch (e) {
-    return {error: e.stack};
-  }
-}
-
-class StringStream extends Readable {
-  constructor(str) {
-    super();
-    this._str = str;
-    this._ended = false;
-  }
-
-  _read() {
-    if (this._ended)
-      return;
-    this._ended = true;
-    process.nextTick(_ => {
-      this.push(Buffer.from(this._str, 'utf8'));
-      this.push(null);
-    });
-  }
-}
-
-async function removeSourceContentIfMatch(sourceMapURL, compiledURL, payload) {
-  const {sourcesContent, sources} = payload;
-  if (!sourcesContent || !sources)
-    return;
-  for (let i = 0; i < sources.length; ++i) {
-    if (!sources[i] || !sourcesContent[i]) continue;
-    let url = sources[i];
-    if (!path.isAbsolute(url))
-      url = path.join(path.dirname(compiledURL), url);
-    if (!fs.existsSync(url))
-      continue;
-    const sourceContentStream = new StringStream(sourcesContent[i]);
-    const sourceContentLines = await readLines(sourceContentStream);
-    const fileStream = fs.createReadStream(url);
-    const fileStreamLines = await readLines(fileStream);
-    if (sourceContentLines.length === fileStreamLines.length) {
-      let equal = true;
-      for (let i = 0; i < sourceContentLines.length; ++i) {
-        if (sourceContentLines[i] !== fileStreamLines[i]) {
-          equal = false;
-          break;
-        }
-      }
-      if (equal)
-        sourcesContent[i] = undefined;
-    }
-  }
-}
-
-async function readLines(stream) {
-  const rl = readline.createInterface({
-    input: stream,
-    crlfDelay: Infinity
-  });
-  return new Promise(resolve => {
-    stream.once('error', _ => resolve(null));
-    const lines = [];
-    rl.on('line', line => lines.push(line));
-    rl.on('close', _ => resolve(lines));
-  });
 }
 
 module.exports = {launch};
