--- test/fs/file1
@@ -1,1 +1,0 @@
-file1
