--- test/fs/folder1/file2
@@ -1,1 +1,0 @@
-file2
