--- test/fs/folder1/folder2/file3
@@ -1,1 +1,0 @@
-file3
