--- test/trivial_service.js
@@ -1,29 +1,0 @@
-/**
- * @license Copyright 2018 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
- */
-
-const {ServiceBase} = require('../services/service_base.js');
-
-class TrivialService extends ServiceBase {
-  sum({a,b}) {
-    return a + b;
-  }
-
-  dispose() {
-    process.exit(0);
-  }
-
-  throwMethod() {
-    throw 42;
-  }
-
-  _getMethod(name) {
-    if (name === 'getMethodError')
-      throw new Error('getMethodError!');
-    return this[name];
-  }
-}
-
-new TrivialService();
